add VideoCacheModule
