+++ app/src/main/java/com/reysl/videoapp/presentation/common/ShimmerEffect.kt
@@ -97,7 +97,7 @@ fun VideoCardShimmerEffect(modifier: Modifier = Modifier) {
 
 
 @Preview(showBackground = true)
-@Preview(showBackground = true, uiMode = Configuration.UI_MODE_NIGHT_YES)
+@Preview(showBackground = true)
 @Composable
 private fun VideoCardShimmerEffectPreview() {
     VideoCardShimmerEffect()
